# spec/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"GenericHealthStatus": {
				"type": "string",
				"enum": [
					"healthy",
					"unhealthy"
				]
			},
			"GenericHealth": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/GenericHealthStatus"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Year": {
				"type": "integer",
				"format": "int32",
				"description": "Common era year"
			},
			"SingleHistoryEvent": {
				"properties": {
					"year": {
						"$ref": "#/components/schemas/Year"
					},
					"data": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"source": {
						"type": "string"
					}
				},
				"required": [
					"year",
					"data",
					"source"
				],
				"type": "object",
				"additionalProperties": {}
			},
			"NotFoundCodeEnum": {
				"type": "string",
				"enum": [
					"NF001",
					"NF002",
					"NF003"
				]
			},
			"NotFoundEvent": {
				"properties": {
					"year": {
						"$ref": "#/components/schemas/Year"
					},
					"body": {
						"type": "string"
					},
					"code": {
						"$ref": "#/components/schemas/NotFoundCodeEnum"
					}
				},
				"required": [
					"year",
					"body",
					"code"
				],
				"type": "object",
				"additionalProperties": {}
			},
			"Time": {
				"type": "string",
				"format": "HH:MM",
				"description": "24-hour clock time",
				"pattern": "^([0-1]?[0-9]|2[0-3]):([0-5][0-9])$"
			},
			"Language": {
				"type": "string",
				"enum": [
					"aa",
					"ab",
					"ace",
					"ady",
					"af",
					"ak",
					"als",
					"alt",
					"am",
					"an",
					"ang",
					"ar",
					"arc",
					"ary",
					"arz",
					"as",
					"ast",
					"atj",
					"av",
					"avk",
					"awa",
					"ay",
					"az",
					"azb",
					"ba",
					"ban",
					"bar",
					"bat-smg",
					"bcl",
					"be",
					"be-tarask",
					"be-x-old",
					"bg",
					"bh",
					"bi",
					"bjn",
					"bm",
					"bn",
					"bo",
					"bpy",
					"br",
					"bs",
					"bug",
					"bxr",
					"ca",
					"cbk-zam",
					"cdo",
					"ce",
					"ceb",
					"ch",
					"cho",
					"chr",
					"chy",
					"ckb",
					"co",
					"cr",
					"crh",
					"cs",
					"csb",
					"cu",
					"cv",
					"cy",
					"da",
					"de",
					"din",
					"diq",
					"dsb",
					"dty",
					"dv",
					"dz",
					"ee",
					"el",
					"eml",
					"en",
					"eo",
					"es",
					"et",
					"eu",
					"ext",
					"fa",
					"ff",
					"fi",
					"fiu-vro",
					"fj",
					"fo",
					"fr",
					"frp",
					"frr",
					"fur",
					"fy",
					"ga",
					"gag",
					"gan",
					"gcr",
					"gd",
					"gl",
					"glk",
					"gn",
					"gom",
					"gor",
					"got",
					"gu",
					"gv",
					"ha",
					"hak",
					"haw",
					"he",
					"hi",
					"hif",
					"ho",
					"hr",
					"hsb",
					"ht",
					"hu",
					"hy",
					"hyw",
					"hz",
					"ia",
					"id",
					"ie",
					"ig",
					"ii",
					"ik",
					"ilo",
					"inh",
					"io",
					"is",
					"it",
					"iu",
					"ja",
					"jam",
					"jbo",
					"jv",
					"ka",
					"kaa",
					"kab",
					"kbd",
					"kbp",
					"kg",
					"ki",
					"kj",
					"kk",
					"kl",
					"km",
					"kn",
					"ko",
					"koi",
					"kr",
					"krc",
					"ks",
					"ksh",
					"ku",
					"kv",
					"kw",
					"ky",
					"la",
					"lad",
					"lb",
					"lbe",
					"lez",
					"lfn",
					"lg",
					"li",
					"lij",
					"lld",
					"lmo",
					"ln",
					"lo",
					"lrc",
					"lt",
					"ltg",
					"lv",
					"mad",
					"mai",
					"map-bms",
					"mdf",
					"mg",
					"mh",
					"mhr",
					"mi",
					"min",
					"mk",
					"ml",
					"mn",
					"mni",
					"mnw",
					"mo",
					"mr",
					"mrj",
					"ms",
					"mt",
					"mus",
					"mwl",
					"my",
					"myv",
					"mzn",
					"na",
					"nah",
					"nap",
					"nds",
					"nds-nl",
					"ne",
					"new",
					"ng",
					"nia",
					"nl",
					"nn",
					"no",
					"nov",
					"nqo",
					"nrm",
					"nso",
					"nv",
					"ny",
					"oc",
					"olo",
					"om",
					"or",
					"os",
					"pa",
					"pag",
					"pam",
					"pap",
					"pcd",
					"pdc",
					"pfl",
					"pi",
					"pih",
					"pl",
					"pms",
					"pnb",
					"pnt",
					"ps",
					"pt",
					"qu",
					"rm",
					"rmy",
					"rn",
					"ro",
					"roa-rup",
					"roa-tara",
					"ru",
					"rue",
					"rw",
					"sa",
					"sah",
					"sat",
					"sc",
					"scn",
					"sco",
					"sd",
					"se",
					"sg",
					"sh",
					"shn",
					"shy",
					"si",
					"simple",
					"sk",
					"skr",
					"sl",
					"sm",
					"smn",
					"sn",
					"so",
					"sq",
					"sr",
					"srn",
					"ss",
					"st",
					"stq",
					"su",
					"sv",
					"sw",
					"szl",
					"szy",
					"ta",
					"tay",
					"tcy",
					"te",
					"tet",
					"tg",
					"th",
					"ti",
					"tk",
					"tl",
					"tn",
					"to",
					"tpi",
					"tr",
					"trv",
					"ts",
					"tt",
					"tum",
					"tw",
					"ty",
					"tyv",
					"udm",
					"ug",
					"uk",
					"ur",
					"uz",
					"ve",
					"vec",
					"vep",
					"vi",
					"vls",
					"vo",
					"wa",
					"war",
					"wo",
					"wuu",
					"xal",
					"xh",
					"xmf",
					"yi",
					"yo",
					"yue",
					"za",
					"zea",
					"zh",
					"zh-classical",
					"zh-min-nan",
					"zh-yue",
					"zu"
				],
				"description": "Language supported by wikipedia"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "wyin-api-feed",
		"version": "0.1.0",
		"description": "REST API serving historical events",
		"license": {
			"name": "MPL-2.0"
		},
		"contact": {
			"name": "mrl5"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/health": {
			"get": {
				"operationId": "GetGenericHealth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericHealth"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/history/event": {
			"get": {
				"operationId": "GetEventByTime",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SingleHistoryEvent"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundEvent"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "t",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Time"
						}
					},
					{
						"in": "query",
						"name": "lang",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Language"
						}
					}
				]
			}
		},
		"/history/event/{year}": {
			"get": {
				"operationId": "GetEventByYear",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SingleHistoryEvent"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundEvent"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "year",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Year"
						}
					},
					{
						"in": "query",
						"name": "lang",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Language"
						}
					}
				]
			}
		},
		"/history/event/random": {
			"get": {
				"operationId": "GetEventByRandom",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SingleHistoryEvent"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundEvent"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "lang",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Language"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}